changes done and the error is debugged its working fine!
